+++ pkg/container/podman_provider.go
@@ -7,7 +7,6 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/cyberark/conjur-inspect/pkg/check"
 	"github.com/cyberark/conjur-inspect/pkg/shell"
 )
 
@@ -17,29 +16,6 @@ var executePodmanInfoFunc = executePodmanInfo
 // PodmanProvider is a concrete implementation of the
 // ContainerProvider interface for Podman
 type PodmanProvider struct{}
-
-type PodmanProviderInfo struct {
-	rawData []byte
-	info    *PodmanInfo
-}
-
-type PodmanInfo struct {
-	Version VersionInfo `json:"version"`
-	Store   StoreInfo   `json:"store"`
-}
-
-// VersionInfo contains the Podman version information
-type VersionInfo struct {
-	Version string
-}
-
-// StoreInfo contains the Podman storage information
-type StoreInfo struct {
-	GraphDriverName string `json:"graphDriverName"`
-	GraphRoot       string `json:"graphRoot"`
-	RunRoot         string `json:"runRoot"`
-	VolumePath      string `json:"volumePath"`
-}
 
 // Name returns the name of the Podman provider
 func (*PodmanProvider) Name() string {
@@ -72,58 +48,9 @@ func (*PodmanProvider) Info() (ContainerProviderInfo, error) {
 	return podmanProviderInfo, nil
 }
 
-func (info *PodmanProviderInfo) Results() []check.Result {
-	return []check.Result{
-		podmanVersionResult(info.info),
-		podmanDriverResult(info.info),
-		podmanGraphRootResult(info.info),
-		podmanRunRootResult(info.info),
-		podmanVolumeRootResult(info.info),
-	}
-}
-
-func (info *PodmanProviderInfo) RawData() []byte {
-	return info.rawData
-}
-
-func podmanVersionResult(podmanInfo *PodmanInfo) check.Result {
-	return check.Result{
-		Title:  "Podman Version",
-		Status: check.StatusInfo,
-		Value:  podmanInfo.Version.Version,
-	}
-}
-
-func podmanDriverResult(podmanInfo *PodmanInfo) check.Result {
-	return check.Result{
-		Title:  "Podman Driver",
-		Status: check.StatusInfo,
-		Value:  podmanInfo.Store.GraphDriverName,
-	}
-}
-
-func podmanGraphRootResult(podmanInfo *PodmanInfo) check.Result {
-	return check.Result{
-		Title:  "Podman Graph Root",
-		Status: check.StatusInfo,
-		Value:  podmanInfo.Store.GraphRoot,
-	}
-}
-
-func podmanRunRootResult(podmanInfo *PodmanInfo) check.Result {
-	return check.Result{
-		Title:  "Podman Run Root",
-		Status: check.StatusInfo,
-		Value:  podmanInfo.Store.RunRoot,
-	}
-}
-
-func podmanVolumeRootResult(podmanInfo *PodmanInfo) check.Result {
-	return check.Result{
-		Title:  "Podman Volume Path",
-		Status: check.StatusInfo,
-		Value:  podmanInfo.Store.VolumePath,
-	}
+// Container returns a Podman container instance for the given ID or name
+func (*PodmanProvider) Container(containerID string) Container {
+	return &PodmanContainer{ContainerID: containerID}
 }
 
 func executePodmanInfo() (stdout, stderr []byte, err error) {